# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API de una Librería -BEDU-",
    "description": "Esta aplicación permite gestionar libros, sus autores, comentarios y usuarios.",
    "contact": {
      "name": "Equipo 3 BEDU",
      "email": "equipo3bedu@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4001/v1",
      "description": "servidor local"
    }
  ],
  "tags": [
    {
      "name": "Libros",
      "description": "Endpoint que maneja toda la información referente a los libros."
    },
    {
      "name": "Autores",
      "description": "Endpoint que maneja toda la información referente a los autores."
    },
    {
      "name": "Usuarios",
      "description": "Endpoint que maneja toda la información referente a los usuarios."
    },
    {
      "name": "Comentarios",
      "description": "Endpoint que maneja toda la información referente a los comentarios."
    }
  ],
  "paths": {
    "/autores": {
      "post": {
        "tags": ["Autores"],
        "summary": "Crear nuevo autor.",
        "operationId": "crearUsuario",
        "responses": {
          "200": {
            "description": "El autor creado"
          }
        },
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto de Autor para crearlo.",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        ]
      },
      "get": {
        "tags": ["Autores"],
        "summary": "Obtener todos los autores",
        "operationId": "getAutor",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Objetos de todos los autores",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        }
      }
    },
    "/autores/{id}": {
      "get": {
        "tags": ["Autores"],
        "summary": "Obtener autor por ID",
        "operationId": "getAutores",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Objeto del autor por su ID",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        }
      },
      "put": {
        "tags": ["Autores"],
        "summary": "Actualizar",
        "operationId": "modificarUsuario",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID del autor"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Pet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "El autor modificado",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          },
          "404": {
            "description": "El autor no existe"
          }
        }
      },
      "delete": {
        "tags": ["Autores"],
        "summary": "Eliminar autor por su ID",
        "operationId": "deleteAutor",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID del autor"
          }
        ],
        "responses": {
          "200": {
            "description": "El autor fue eliminado"
          }
        }
      }
    },
    "/Libros": {
      "get": {
        "tags": ["Libros"],
        "summary": "Obtiene todos los libros",
        "description": "Regresa un array de objetos con todos los libros encontrados en la base de datos.",
        "responses": {
          "200": {
            "description": "Solicitud de libros exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLibrosAll"
                }
              }
            }
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      },
      "post": {
        "tags": ["Libros"],
        "summary": "Inserta un nuevo libro a la BD",
        "description": "Crea un nuevo libro con sus datos y lo inserta en la base de datos.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyLibrosPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Libro ingresado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExitoIngresoLibro"
                }
              }
            }
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      }
    },
    "/Libros/{idLibro}": {
      "get": {
        "tags": ["Libros"],
        "summary": "Búsqueda de un único libro",
        "description": "Regresa el libro que coincida con el ID",
        "parameters": [
          {
            "name": "idLibro",
            "in": "path",
            "description": "Id del libro",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLibrosAll"
                }
              }
            }
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      },
      "put": {
        "tags": ["Libros"],
        "summary": "Actualiza un libro",
        "description": "Busca un libro por su ID y actualiza los datos enviados en el body.",
        "parameters": [
          {
            "name": "idLibro",
            "in": "path",
            "description": "Id del libro",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyLibrosPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Libro actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExitoLibrosPut"
                }
              }
            }
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      },
      "delete": {
        "tags": ["Libros"],
        "summary": "Elimina un libro de la BD.",
        "description": "Busca un libro por su ID y lo elimina de la base de datos.",
        "parameters": [
          {
            "name": "idLibro",
            "in": "path",
            "description": "Id del libro",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Libro eliminado exitosamente"
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      }
    },
    "/Libros/Titulo/{frase}": {
      "get": {
        "tags": ["Libros"],
        "summary": "Búsqueda de libros según la frase ingresada",
        "description": "Regresa el (los) libro (s) que coincidan en el título con la palabra o frase ingresada",
        "parameters": [
          {
            "name": "frase",
            "in": "path",
            "description": "Palabra(s) o título completo",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLibrosAll"
                }
              }
            }
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      }
    },
    "/Libros/MasComentados": {
      "get": {
        "tags": ["Libros"],
        "summary": "Muestra lista descendente de libros más comentados.",
        "description": "Regresa un listado ordenado de mayor a menor con los libros más comentados, desde un libro hasta n cantidad, según se envíe en \"?cantidad=n\"",
        "parameters": [
          {
            "name": "cantidad",
            "in": "query",
            "description": "Cantidad de libros a mostrar de los más comentados",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMoreCommentsBooks"
                }
              }
            }
          },
          "400": {
            "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
          },
          "401": {
            "description": "(Unauthorized) No hay autorización para llamar el servicio"
          },
          "404": {
            "description": "(NotFound) No se encontró información"
          },
          "500": {
            "description": "(ServerError) Error en el servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Libro": {
        "required": ["idAuthors", "isbn", "title"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "title": {
            "type": "string"
          },
          "idAuthors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Libro_idAuthors"
            }
          },
          "isbn": {
            "type": "string"
          },
          "isbn13": {
            "type": "string"
          },
          "publication_year": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "pages": {
            "type": "integer"
          },
          "price": {
            "type": "integer"
          },
          "image_url": {
            "type": "string"
          },
          "small_image_url": {
            "type": "string"
          }
        }
      },
      "BodyLibrosPost": {
        "required": ["idAuthors", "isbn", "title"],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Aprendiendo NodeJS en BEDU"
          },
          "idAuthors": {
            "type": "array",
            "items": {
              "example": "613d5cf346ed939f5bfe3d7d, 613d5cf346ed939f5bfe3d7d"
            }
          },
          "isbn": {
            "type": "string",
            "example": "307397674"
          }
        }
      },
      "ExitoIngresoLibro": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "613d8a2546ed939f5bfe3e2f"
          },
          "title": {
            "type": "string",
            "example": "Aprendiendo NodeJS en BEDU"
          },
          "idAuthors": {
            "type": "array",
            "items": {
              "example": "613d5cf346ed939f5bfe3d7d, 613d5cf346ed939f5bfe3d7d"
            }
          },
          "isbn": {
            "type": "string",
            "example": "307397674"
          },
          "isbn13": {
            "type": "string"
          },
          "publication_year": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "pages": {
            "type": "integer"
          },
          "price": {
            "type": "integer"
          },
          "image_url": {
            "type": "string"
          },
          "small_image_url": {
            "type": "string"
          }
        }
      },
      "GetLibrosAll": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/GetLibrosAll_inner"
        }
      },
      "BodyLibrosPut": {
        "type": "object",
        "properties": {
          "isbn13": {
            "type": "string",
            "example": "9780307277670"
          },
          "publication_year": {
            "type": "string",
            "example": "2003"
          },
          "language": {
            "type": "string",
            "example": "eng"
          },
          "publisher": {
            "type": "string",
            "example": "Arcade Publishing"
          },
          "pages": {
            "type": "integer",
            "example": 1078
          },
          "price": {
            "type": "integer",
            "example": 10
          },
          "image_url": {
            "type": "string",
            "example": "https://images.gr-assets.com/books/1303252999m/968.jpg"
          },
          "small_image_url": {
            "type": "string",
            "example": "https://images.gr-assets.com/books/1303252999s/968.jpg"
          }
        }
      },
      "ExitoLibrosPut": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "613d8a2546ed939f5bfe3e2f"
          },
          "title": {
            "type": "string",
            "example": "Aprendiendo NodeJS en BEDU"
          },
          "idAuthors": {
            "type": "array",
            "items": {
              "example": "613d5cf346ed939f5bfe3d7d, 613d5cf346ed939f5bfe3d7d"
            }
          },
          "isbn": {
            "type": "string",
            "example": "307397674"
          },
          "isbn13": {
            "type": "string",
            "example": "9780307277670"
          },
          "publication_year": {
            "type": "string",
            "example": "2003"
          },
          "language": {
            "type": "string",
            "example": "eng"
          },
          "publisher": {
            "type": "string",
            "example": "Arcade Publishing"
          },
          "pages": {
            "type": "integer",
            "example": 1078
          },
          "price": {
            "type": "integer",
            "example": 10
          },
          "image_url": {
            "type": "string",
            "example": "https://images.gr-assets.com/books/1303252999m/968.jpg"
          },
          "small_image_url": {
            "type": "string",
            "example": "https://images.gr-assets.com/books/1303252999s/968.jpg"
          }
        }
      },
      "GetMoreCommentsBooks": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/GetMoreCommentsBooks_inner"
        }
      },
      "Libro_idAuthors": {
        "properties": {
          "author_1": {
            "type": "string"
          },
          "author_n": {
            "type": "string"
          }
        }
      },
      "GetLibrosAll_inner": {
        "properties": {
          "_id": {
            "type": "string",
            "example": "613d8a2546ed939f5bfe3e42"
          },
          "title": {
            "type": "string",
            "example": "The Da Vinci Code"
          },
          "isbn": {
            "type": "string",
            "example": "307277674"
          },
          "isbn13": {
            "type": "string",
            "example": "9780307277670"
          },
          "publication_year": {
            "type": "string",
            "example": "2003"
          },
          "language": {
            "type": "string",
            "example": "eng"
          },
          "publisher": {
            "type": "string",
            "example": "Arcade Publishing"
          },
          "pages": {
            "type": "integer",
            "example": 1078
          },
          "price": {
            "type": "integer",
            "example": 10
          },
          "image_url": {
            "type": "string",
            "example": "https://images.gr-assets.com/books/1303252999m/968.jpg"
          },
          "small_image_url": {
            "type": "string",
            "example": "https://images.gr-assets.com/books/1303252999s/968.jpg"
          },
          "Autores": {
            "type": "object",
            "properties": {
              "_idAuthor": {
                "type": "string",
                "example": "613d5cf346ed939f5bfe3d7d"
              },
              "nombre_completo": {
                "type": "string",
                "example": "Dan Brown"
              }
            }
          }
        }
      },
      "GetMoreCommentsBooks_inner": {
        "properties": {
          "title": {
            "type": "string",
            "example": "The Da Vinci Code"
          },
          "comments": {
            "type": "array",
            "items": {
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Louis"
                },
                "texto": {
                  "type": "string",
                  "example": "Fascinante, te atrapa desde la primer página. El profesor Langdon es el héroe menos esperado."
                },
                "fecha": {
                  "type": "string",
                  "example": "2003-11-24T06:00:00.000Z"
                }
              }
            }
          },
          "TotalComments": {
            "type": "integer",
            "example": 1
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "(Unauthorized) No hay autorización para llamar el servicio"
      },
      "NotFound": {
        "description": "(NotFound) No se encontró información"
      },
      "BadRequest": {
        "description": "(BadRequest) Datos enviados incorrectos o hay datos obligatorios no enviados"
      },
      "ServerError": {
        "description": "(ServerError) Error en el servidor"
      }
    },
    "parameters": {
      "token": {
        "name": "token",
        "in": "header",
        "description": "Token de autenticación.",
        "required": true,
        "style": "simple",
        "explode": false,
        "schema": {
          "type": "string"
        }
      },
      "idLibro": {
        "name": "idLibro",
        "in": "path",
        "description": "Id del libro",
        "required": true,
        "style": "simple",
        "explode": false,
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "definitions": {
    "Autor": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "idBooks": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Libro": {
      "type": "object",
      "required": ["title", "idAuthors", "isbn"],
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "idAuthors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "isbn": {
          "type": "string"
        },
        "isbn13": {
          "type": "string"
        },
        "publication_year": {
          "type": "string"
        },
        "language": {
          "type": "string"
        },
        "publisher": {
          "type": "string"
        },
        "pages": {
          "type": "integer"
        },
        "price": {
          "type": "number"
        },
        "image_url": {
          "type": "string"
        },
        "small_image_url": {
          "type": "string"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "required": ["username", "email", "nombre", "apellido"],
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "tipo": {
          "type": "string",
          "enum": ["admin", "cliente"]
        }
      }
    },
    "Comentario": {
      "type": "object",
      "required": ["nombre", "email", "texto"],
      "properties": {
        "id": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "libro_id": {
          "type": "string"
        },
        "texto": {
          "type": "string"
        },
        "fecha": {
          "type": "string"
        }
      }
    }
  }
}